[refactor] #123 responsebody package 내 chat package 생성
